+++ src/class.rs
@@ -40,7 +40,7 @@ pub struct MidiClass<'a, B: UsbBus> {
 
 /// Error variants for read operations.
 #[derive(Debug, Clone, Eq, PartialEq)]
-pub enum MidiReadError {
+pub enum UsbMidiReadError {
     /// Parsing of the packet failed.
     ParsingFailed(UsbMidiEventPacketError),
     /// USB stack error returned from `usb-device`.